Found a solution to the problem when compiling in unix.
In unix, if you compile from outside the src file - you get package
issues with the junit tests because you have to set the class path to
the junit class, rather than the src. 

I didn't find a way to have two class paths at once, so I instead hard 
coded a filter in the tests. When the tests obtain the current path,
they check if it is in the src file -- and if the tests remove /src from
the path so that tests will be consistent and can find the TestData/.sb2
files. This occured in last commit, rather than this one. But that
message was in complete so I quickly did another commit