Re-designed how GUI configuration for graders will be done.
 Now there is a popup-dialog class called ConfigureDialog. This class
wraps a customized panel that is used to set configuratin parameters.
The ConfigureDialog class also encompasses the OK and CANCEL behavior so
that those features are closed for modification and can be trusted to
work by the GUI programmers. Also, in order to make a customizeable gui
window (say for sprite grader), the programmer now creates a class that
extends ConfigurablePanel. This abstract class requires the creation of
certain methods for the ConfigurableDialog to successfully do its work
(such as OKPressed and CancelPressed having an impact on the
configuration parameters).